Update REST API

+++ src/main/java/com/aceba1/cd/capstone/processor/service/TransactionService.java
@@ -77,8 +77,10 @@ public class TransactionService {
     repository.delete(transaction);
   }
 
-  public void deleteById(Long id) {
+  public boolean deleteById(Long id) {
+    if (repository.findById(id).isEmpty()) return false;
     repository.deleteById(id);
+    return true;
   }
 
   public long getSize() {
